part payment done
